NVMeBasicContext: Properly cleanup resources, allowing destruction
nvmesensor was terminating with an uncaught exception, e.g:

    Jun 03 04:52:09 bmc nvmesensor[507]: terminate called after throwing an instance of 'sdbusplus::exception::SdBusError'
    Jun 03 04:52:09 bmc nvmesensor[507]:   what():  sd_bus_add_object_vtable: org.freedesktop.DBus.Error.FileExists: File exists

This would occur whenever entity-manager published a configuration for a
new drive in the system. The implementation of nvmesensor isn't the
smartest, and it tries to just scrap all NVMeContexts it knows of, along
with their NVMeSensor instances, and reconstruct them all from newly
captured entity-manager configuration data.

The problem lies in the fact that the NVMeContexts were not getting
destructed due embedding of std::shared_ptrs obtained via
shared_from_this() into async callback lambda captures whose context was
owned by the NVMeContext implementation.

This patch untangles the relationships by removing
std::enable_shared_from_this from the inheritance hierarchy for
NVMeContext and switching the context map to instead use
std::unique_ptr<NVMeContext>.

With this we are able to successfully destruct the NVMeContext
derivative instances via the clear() method on the context map.

However, there's more to the story, as the NVMeSensors owned by a given
context are asynchronously iterated for polling purposes. To make sure
we don't have races from unbounded latency on destruction of NVMeSensor
instances due to asynchronous IO (and thus make sure the creation of the
replacement DBus objects won't fail) we define a destructor for
NVMeBasicContext that cleans up the async streams associated with the
NVMeBasicContext's IO thread, freeing the active NVMeSensor instance by
immediately invoking the callback with an error-code, and forcing
termination of the IO thread through pipe access failures. We then join
the thread from the constructor to complete the cleanup.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Change-Id: I63d0e7eb3318c209b08551072a3cb7279da21269