psusensor: don't add fan sensor if not in labels
The checkPWMSensor() function will add a fan sensor to the global array
even if the user has not requested it be added. This has caused issues
on certain systems like IBM system1 where the device driver does not
fully support the hwmon requirements for fan speeds. This results in the
GUI showing the PSU fan sensors but them not having any valid value.

This commit duplicates the label check logic a little bit further down
in the code so it's not optimal. But without completely refactoring this
huge function, this seems like the best option.

The current bug is that checkPWMSensor() is called and the sensor is
added to pwmSensors[]. Then once we return from the call, the label
check is done and a continue is issued because the fan sensor is not
in the label list. Because of this, none of the additional code is
run required to initialize the sensor properly.

Tested
- Confirmed the fan sensors are no longer created on IBM System1

Change-Id: If801981295f36cc845e559a76573b488e9c420ac
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>